+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Other
 - Refactored train script, now uses a `ExperimentManager` class
+- Replaced `make_env` with SB3 built-in `make_vec_env`
 
 ## Pre-Release 0.10.0 (2020-10-28)